Reorganise frontend components
The directory structure here made it a bit confusing to jump from a
top-level exported component to its source, which could be deeply tucked
away in a non-obvious spot. We also had some exports that were not
obviously named according to their corresponding feature.
