+++ .changeset/clever-pillows-laugh.md
@@ -1,0 +1,5 @@
+---
+'wingman-fe': minor
+---
+
+**QuestionnaireQueryOutput:** Export component
