+++ .changeset/empty-singers-heal.md
@@ -1,0 +1,13 @@
+---
+'wingman-fe': minor
+---
+
+**QuestionnaireCreateInput:** Remove export
+
+A similar type can be derived from component props:
+
+```typescript
+import {ComponentProps} from 'react';
+
+type Input = ComponentProps<typeof QuestionnaireBuilder>['graphqlInput']
+```
