+++ .changeset/perfect-hornets-mate.md
@@ -1,0 +1,5 @@
+---
+'wingman-fe': minor
+---
+
+**FormBuilderQueryInput:** Rename to `QuestionnaireQueryInput`
