--- fe/lib/components/Questionnaire/index.ts
@@ -1,10 +1,0 @@
-export { GraphqlQueryRenderer } from './components/GraphqlQueryRenderer/GraphqlQueryRenderer';
-export type { QuestionnaireCreateInput } from './components/GraphqlQueryRenderer/GraphqlQueryRenderer';
-
-export type {
-  FormComponent,
-  FreeTextQuestion,
-  PrivacyConsent,
-  Question,
-  SelectionQuestion,
-} from './types';
