+++ fe/lib/components/QuestionnaireBuilder/FormBuilder/components/NewQuestionForm/components/DisplayOption.tsx
