+++ fe/lib/components/QuestionnaireBuilder/FormBuilder/components/NewQuestionForm/components/QuestionInputFields.tsx
@@ -1,7 +1,10 @@
 import { Stack, TextField } from 'braid-design-system';
 import React from 'react';
 
-import type { QuestionType, ResponseChoice } from '../../../../../types';
+import type {
+  QuestionType,
+  ResponseChoice,
+} from '../../../../../../private/questionnaires/types';
 
 import SelectOptions from './SelectOptions';
