--- fe/lib/components/index.ts
@@ -1,9 +1,0 @@
-export { JobCategorySelect } from './JobCategorySelect/JobCategorySelect';
-export { JobCategorySuggest } from './JobCategorySuggest/JobCategorySuggest';
-export { LocationSuggest } from './LocationSuggest/LocationSuggest';
-
-export { SpecifiedPersonForm } from './Person/SpecifiedPersonForm';
-export { QuestionnaireBuilder } from './Questionnaire/QuestionnaireBuilder/QuestionnaireBuilder';
-export { QuestionnaireForm } from './Questionnaire/QuestionnaireForm/QuestionnaireForm';
-export type { QuestionnaireCreateInput } from './Questionnaire/components/GraphqlQueryRenderer/GraphqlQueryRenderer';
-export { FormBuilderQueryInput } from './Questionnaire/components/FormBuilderQueryInput';
