+++ fe/lib/index.ts
@@ -1,1 +1,11 @@
-export * from './components';
+export { JobCategorySelect } from './components/JobCategorySelect/JobCategorySelect';
+export { JobCategorySuggest } from './components/JobCategorySuggest/JobCategorySuggest';
+
+export { LocationSuggest } from './components/LocationSuggest/LocationSuggest';
+
+export { QuestionnaireBuilder } from './components/QuestionnaireBuilder/QuestionnaireBuilder';
+export { QuestionnaireForm } from './components/QuestionnaireForm/QuestionnaireForm';
+export { QuestionnaireQueryInput } from './components/QuestionnaireQueryInput/QuestionnaireQueryInput';
+export { QuestionnaireQueryOutput } from './components/QuestionnaireQueryOutput/QuestionnaireQueryOutput';
+
+export { SpecifiedPersonForm } from './components/SpecifiedPersonForm/SpecifiedPersonForm';
