+++ fe/lib/scoobie.d.ts
